requirements.txt
google-generativeai>=0.8.3
python-dotenv>=1.0